ReadDataIn function added
The read_data_in function takes data from the DRAM and processes it, then saves it to the SD card. Just as the PrintData function, this function reads continuously until the 'q' character is received from the user. The data are saved in a binary file on the SD card and may be processed using my bin_to_text program located on my shared drive space. This is version 1.1 of the devkit.